fix: remove more redundant u32 casts on mozjpeg_sys JDIMENSION

+++ zenjpeg/tests/diagnose_decoder_diff.rs
@@ -265,11 +265,11 @@ fn decode_mozjpeg_rgb(data: &[u8]) -> (u32, u32, Vec<u8>) {
         assert_eq!(jpeg_read_header(&mut ci, 1), 1);
         ci.out_color_space = J_COLOR_SPACE::JCS_RGB;
         jpeg_start_decompress(&mut ci);
-        let w = ci.output_width as u32;
-        let h = ci.output_height as u32;
+        let w = ci.output_width;
+        let h = ci.output_height;
         let stride = w as usize * ci.output_components as usize;
         let mut out = vec![0u8; h as usize * stride];
-        while (ci.output_scanline as u32) < h {
+        while ci.output_scanline < h {
             let off = ci.output_scanline as usize * stride;
             let mut p = out[off..].as_mut_ptr();
             jpeg_read_scanlines(&mut ci, &mut p, 1);
@@ -357,11 +357,11 @@ fn decode_mozjpeg_ycbcr(data: &[u8]) -> (u32, u32, Vec<u8>) {
         assert_eq!(jpeg_read_header(&mut ci, 1), 1);
         ci.out_color_space = J_COLOR_SPACE::JCS_YCbCr;
         jpeg_start_decompress(&mut ci);
-        let w = ci.output_width as u32;
-        let h = ci.output_height as u32;
+        let w = ci.output_width;
+        let h = ci.output_height;
         let stride = w as usize * ci.output_components as usize;
         let mut out = vec![0u8; h as usize * stride];
-        while (ci.output_scanline as u32) < h {
+        while ci.output_scanline < h {
             let off = ci.output_scanline as usize * stride;
             let mut p = out[off..].as_mut_ptr();
             jpeg_read_scanlines(&mut ci, &mut p, 1);
